amelioration du css du header, du footer, nommage des h1 des composants pour les pages légales du site
