+++ src/components/Header/header.scss
@@ -17,8 +17,6 @@
   background-color: $light-color;
 }
 
-// .header-border {
-//   border-bottom: 3px solid $secondary-color;
-//   margin: none;
-//   padding: none;
-// }
+.header-shadow {
+  box-shadow: 0px 10px 5px -4px $secondary-color;
+}
