+++ src/pages/Cookies/Cookies.jsx
@@ -2,10 +2,11 @@ import React from "react";
 
 const Cookies = () => {
   return (
-    <div className="container">
-      <h1>Cookies</h1>
-    </div>
-  )
-}
+    <main className="container">
+      <h1 className="m-5">Cookies</h1>
+      <p>A venir (en cours de production)</p>
+    </main>
+  );
+};
 
-export default Cookies
+export default Cookies;
