+++ src/pages/Données_Perso/Donnees.jsx
@@ -2,10 +2,11 @@ import React from "react";
 
 const Donnees = () => {
   return (
-    <div className="container">
-      <h1>Données Personnelles</h1>
-    </div>
-  )
-}
+    <main className="container">
+      <h1 className="m-5">Données Personnelles</h1>
+      <p>A venir (en cours de production)</p>
+    </main>
+  );
+};
 
-export default Donnees
+export default Donnees;
