+++ src/pages/Legal/Legal.jsx
@@ -2,9 +2,10 @@ import React from "react";
 
 const Legal = () => {
   return (
-    <div className="container">
-      <h1>Mentions Légales</h1>
-    </div>
+    <main className="container">
+      <h1 className="m-5">Mentions Légales</h1>
+      <p>A venir (en cours de production)</p>
+    </main>
   );
 };
