+++ src/pages/NotFound/NotFound.jsx
@@ -4,18 +4,18 @@ import img404 from "../../assets/images/error-404.svg";
 
 const NotFound = () => {
   return (
-    <div className="container body-wrapper my-5">
-      <h1>Erreur 404</h1>
+    <main className="container body-wrapper my-5">
+      <h1 className="text-center">Erreur 404 - Page non trouvé</h1>
       <div className="text-center">
         <img src={img404} alt='logo ordinateur affichant "404"' width="512" height="512"/>
       </div>
-      <p>Page non trouvée...</p>
       <Button
         linkPage="/"
         type="button"
         buttonName="Retourner à la page d'accueil"
+        color="blue"
       />
-    </div>
+    </main>
   );
 };
